ROC-3661: extracted partialAdmissionGuard
